+++ src/internal/packager/helm/repo.go
@@ -35,8 +35,8 @@ import (
 	"github.com/zarf-dev/zarf/src/config"
 	"github.com/zarf-dev/zarf/src/config/lang"
 	"github.com/zarf-dev/zarf/src/internal/git"
+	"github.com/zarf-dev/zarf/src/pkg/ocitransport"
 	"github.com/zarf-dev/zarf/src/pkg/transform"
-	negotiate "github.com/zarf-dev/zarf/src/pkg/transport"
 	"github.com/zarf-dev/zarf/src/pkg/utils"
 )
 
@@ -53,7 +53,7 @@ func negotiateChartPlainHTTP(ctx context.Context, ociURL string, remoteOptions t
 	if err != nil {
 		return false, fmt.Errorf("unable to parse chart url %q: %w", ociURL, err)
 	}
-	plainHTTP, err := negotiate.From(ctx).UsePlainHTTP(ctx, ref.Registry, negotiate.ProbeOptions{InsecureSkipTLSVerify: remoteOptions.InsecureSkipTLSVerify})
+	plainHTTP, err := ocitransport.From(ctx).UsePlainHTTP(ctx, ref.Registry, ocitransport.ProbeOptions{InsecureSkipTLSVerify: remoteOptions.InsecureSkipTLSVerify})
 	if err != nil {
 		return false, fmt.Errorf("unable to reach chart registry for %q: %w", ociURL, err)
 	}
