+++ src/pkg/ocitransport/ocitransport_test.go
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
 // SPDX-FileCopyrightText: 2021-Present The Zarf Authors
 
-package transport
+package ocitransport
 
 import (
 	"context"
